+++ inst/app/R/server_global.R
@@ -82,9 +82,9 @@ server_global <- function(input, output, session, rv, config_dir, gcfg_rv) {
       )
     )
     tryCatch({
-      write_global_config(gcfg, global_config_path(config_dir()))
-      register_report_resource_path(gcfg$report_output_dir, config_dir())
-      gcfg_rv(gcfg)
+      merged <- update_global_config(gcfg, global_config_path(config_dir()))
+      register_report_resource_path(merged$report_output_dir, config_dir())
+      gcfg_rv(merged)
       showNotification("Global config saved.", type="message", duration=3)
     }, error = function(e) {
       showModal(modalDialog(